fixed example test for pulsarpp
